requirements.txt
# ArcHillx v1.0.0 — Python dependencies
# Install: pip install -r requirements.txt

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Settings / config ─────────────────────────────────────────────────────────
pydantic-settings>=2.2.0
python-dotenv>=1.0.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.29
# MySQL support (install when DB_TYPE=mysql)
pymysql>=1.1.0
# MSSQL support (install when DB_TYPE=mssql; requires ODBC Driver 17 on host)
# pyodbc>=5.0.0

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler>=3.10.4

# ── YAML ──────────────────────────────────────────────────────────────────────
pyyaml>=6.0.1

# ── AI Providers (install only what you need) ─────────────────────────────────
anthropic>=0.28.0
openai>=1.30.0
google-generativeai>=0.7.0

# ── Skills ────────────────────────────────────────────────────────────────────
# Web search (choose one or both)
duckduckgo-search>=6.0.0       # ddgs CLI + Python API
# playwright>=1.44.0           # optional: headless browser automation

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx>=0.27.0                   # async HTTP client (used by some providers)

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0

# ── Migrations ─────────────────────────────────────────────────────────────────
alembic>=1.13.0
